Adding templates class
